Telegram f1 standing/requirements.txt
python-telegram-bot==13.7
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
Flask==2.0.1
requests==2.26.0
python-dotenv==0.19.0 